backend/requirements.txt
flask>=2.3.0
paho-mqtt>=1.6.1
requests>=2.31.0
